Fikser skiplinks og kb-navigering. Diverse refaktorering.
