+++ src/komponenter/header/header-regular/HeaderMenylinje.tsx
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux';
 import { useKbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
 import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 import { GACategory } from 'utils/google-analytics';
-import { SokDropdown } from 'komponenter/header/header-regular/desktop/sok/SokDropdown';
+import { SokDropdown } from 'komponenter/header/header-regular/desktop/sok-dropdown/SokDropdown';
 import { MinsideMeny } from 'komponenter/header/header-regular/desktop/minside-meny/MinsideMeny';
 import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
 import { VarslerDropdown } from 'komponenter/header/header-regular/varsler/VarslerDropdown';
@@ -15,6 +15,7 @@ import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper
 import { SokKnapp } from 'komponenter/header/header-regular/common/knapper/sok-knapp/SokKnapp';
 import { MinsideKnapp } from 'komponenter/header/header-regular/common/knapper/minside-knapper/MinsideKnapp';
 import { VarslerKnapp } from 'komponenter/header/header-regular/common/knapper/varsler-knapp/VarslerKnapp';
+import 'komponenter/header/header-regular/HeaderMenylinje.less';
 
 export const headerLogoId = 'header-logo-id';
 export const desktopHovedmenyKnappId = 'desktop-hovedmeny-knapp-id';
