+++ src/komponenter/header/header-regular/desktop/sok-dropdown/SokDropdown.less
