+++ src/komponenter/header/header-regular/skiplinks/Skiplinks.less
@@ -1,4 +1,4 @@
-@import (reference) '../../../styling-variabler.less';
+@import (reference) '../../../../styling-variabler.less';
 
 .site-skiplinks {
     ul {
