+++ src/komponenter/header/header-regular/skiplinks/Skiplinks.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
-import { createStore } from '../../../store';
+import { createStore } from 'store';
 import { Provider as ReduxProvider } from 'react-redux';
 import Skiplinks from './Skiplinks';
